+++ x/precisebank/keeper/keeper.go
@@ -1,0 +1,23 @@
+package keeper
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
+)
+
+// TODO: Enforce that Keeper implements the expected keeper interfaces
+// var _ types.BankKeeper = Keeper{}
+
+// Keeper defines the precisebank module's keeper
+type Keeper struct {
+	cdc      codec.BinaryCodec
+	storeKey storetypes.StoreKey
+}
+
+// NewKeeper creates a new keeper
+func NewKeeper(cdc codec.BinaryCodec, storeKey storetypes.StoreKey) Keeper {
+	return Keeper{
+		cdc:      cdc,
+		storeKey: storeKey,
+	}
+}